Rename AnalyticsIntegrationClient
